requirements.txt
streamlit>=1.28.0
groq>=0.4.0
python-dotenv>=1.0.0
Pillow>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0
pandas>=2.0.0
tensorflow>=2.13.0
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.0
scikit-learn>=1.3.0
seaborn>=0.12.0
tqdm>=4.65.0
httpx>=0.24.0
lime
scikit-image